pip_requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
debugpy==1.4.1
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
importlib-metadata==5.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.0
joblib==1.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lpips==0.1.4
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.2
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.0
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging==21.3
pandas==1.5.3
pandocfilters==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
requests-oauthlib==1.3.0
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
tabulate==0.9.0
tensorboard-data-server==0.6.1
tensorboardX==2.4
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.7.1
torchattacks==3.3.0
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado==6.1
tqdm==4.64.1
traitlets==5.5.0
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.3
